Generate a parameter label combining specimen type, parameter name, and units

Generate a parameter label combining specimen type, parameter name, and units

Usage

make_param_label(
  spec,
  param,
  unit,
  expect_single = TRUE,
  allow_missing_spec = TRUE,
  allow_missing_unit = TRUE
)

Arguments

spec

Specimen type (typically, "blood", "plasma", etc.)

param

Parameter name (the thing that was measured)

unit

Units of measure

expect_single

Should a single parameter label be generated (or may it be a vector)?

allow_missing_spec, allow_missing_unit

Are the spec or unit parameters allowed to be NA?

Value

A character vector.

See also