Merge a supplementary dataset into a primary dataset

Merge a supplementary dataset into a primary dataset

Usage

merge_supp(
  primary,
  supplementary,
  remove_attributes = c("label"),
  auto_convert = FALSE
)

Arguments

primary

the data.frame of the primary dataset

supplementary

the data.frame of the supplementary dataset

remove_attributes

If TRUE, remove all attributes from all columns (this will break many classes); if FALSE, remove no attributes from any columns; if a character, remove those attributes from the columns.

auto_convert

should the data be automatically converted using `type_convert()`?

Value

primary merged with supplementary where new column names come from the QNAM column in supplementary

See also