Build an ISO8601 date pattern from the year, middle (month or week), day, and time patterns.

Build an ISO8601 date pattern from the year, middle (month or week), day, and time patterns.

Usage

pattern_ISO8601_date_builder(
  truncated,
  pattern_year,
  pattern_middle,
  pattern_day,
  pattern_time,
  allow_before_year_1583
)

Arguments

truncated

Should the date/time be allowed to be truncated? An integer indicating the highest required precision (0=second is required, 1=minute, 2=hour, 3=day, 4=month or week, 5=year, 6=none). A value of 6 will allow an empty string to match.

pattern_year

The pattern for the ISO8601 year (ignored if allow_before_year_1583 is given); if null, only the middle and day part will be returned without a year part.

pattern_middle

The pattern for the ISO8601 month or week (or NULL for ordinal dates)

pattern_day

The pattern for the ISO8601 day

pattern_time

A string to add to the day for including time with date (see pattern_ISO8601_time())

allow_before_year_1583

Should years between 0 and 1582 be allowed (they are only allowed in ISO 8601 with mutual agreement)